+++ admin/dashboard/fetchcommission.php
@@ -7,8 +7,10 @@ if (isset($_GET['month']) && is_numeric($_GET['month'])) {
     $month = (int)$_GET['month']; // Sanitize the month parameter
     $year = date("Y"); // Use the current year by default. Adjust if needed.
 
-    // Query to retrieve order counts grouped by date for the selected month
-    $query = "SELECT DATE(purchase_date) AS order_date, COUNT(*) AS order_count 
+    // Query to retrieve order counts and total commission grouped by date for the selected month
+    $query = "SELECT DATE(purchase_date) AS order_date, 
+                     COUNT(*) AS order_count, 
+                     SUM(commission) AS total_commission 
               FROM orders 
               WHERE MONTH(purchase_date) = ? AND YEAR(purchase_date) = ? 
               GROUP BY DATE(purchase_date) 
@@ -22,15 +24,21 @@ if (isset($_GET['month']) && is_numeric($_GET['month'])) {
         // Fetch data
         $data = [];
         while ($row = $result->fetch_assoc()) {
-            $data[] = $row;
+            $data[] = [
+                "order_date" => $row["order_date"],
+                "order_count" => (int)$row["order_count"], // Convert to integer for consistency
+                "total_commission" => (float)$row["total_commission"] // Convert to float for consistency
+            ];
         }
 
         // Return data as JSON
         echo json_encode($data);
     } else {
+        // Error in preparing the query
         echo json_encode(["error" => "Failed to prepare query"]);
     }
 } else {
+    // Missing or invalid month parameter
     echo json_encode(["error" => "Invalid or missing month parameter"]);
 }
 ?>
